requirements.txt
# =============================================================================
# REQUIREMENTS - RAG CHATBOT VĂN HÓA TÀY & BA NA
# =============================================================================
# Tương thích với Python 3.9 - 3.13
# =============================================================================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Google AI / Gemini (Package MỚI - thay thế google-generativeai)
google-genai>=1.0.0

# Vector Database
chromadb>=0.5.0

# Utilities
python-dotenv>=1.0.0

# Optional: For production
gunicorn>=21.2.0
